+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 * **Breaking change**: Refactored [HexaPDF::Layout::Frame] and associated data
   structures so that the complete result of fitting a box is returned
+* [HexaPDF::Layout::Frame] to use a better algorithm for trimming the shape
 
 ### Fixed